First commit
This should have been an easy assignment but I've been tearing my hair out for days. I seriously thought that I was too dumb to do it for a while. I felt like I understood it in my head but nothing worked how it was supposed to. After several days of that weird stuff started happening when I would try to load the file. Turns out there was something wrong with it, like it was corrupted or something. All that time I thought I was doing it wrong!  A learning experience to be sure.